refactor: 인터셉터를 활용해 controller 메서드 실행 전에 로깅하여 코드 공통화
